fix problems displaying current partition table

+++ fdisk_screen.c
@@ -41,8 +41,8 @@ void screen_hex(unsigned int addr,long value)
   POKE(addr+7,to_screen_hex(value>>0));
 }
 
-void format_hex(const int addr,const int value, const char columns)
-{
+void format_hex(const int addr,const long value, const char columns)
+{  
   char i,c;
   char dec[9];
   screen_hex((int)&dec[0],value);
