+++ fdisk_screen.h
@@ -53,7 +53,7 @@ void recolour_last_line(char colour);
 char read_line(char *buffer,char maxlen);
 
 void format_decimal(const int addr,const int value, const char columns);
-void format_hex(const int addr,const int value, const char columns);
+void format_hex(const int addr,const long value, const char columns);
 
 extern long screen_line_address;
